backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-auth-adfs>=1.9.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
requests>=2.31.0
gunicorn>=21.2.0
whitenoise>=6.6.0
drf-yasg>=1.21.7  # For Swagger/OpenAPI documentation
djangorestframework-simplejwt==5.5.0 